backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
PyJWT==2.10.1
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
pytest==8.3.4
